# TennisPostgresWebApiCSharp/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "TennisPostgresWebApiCSharp",
    "version": "1.0"
  },
  "paths": {
...
    "/api/Players": {
      "get": {
        "tags": [
          "Players"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Player"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Player"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Player"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Players"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Player"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/Player"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/Player"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/Player"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Player"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Player"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/Players/{id}": {
      "get": {
        "tags": [
          "Players"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/Player"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Player"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Player"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Players"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Player"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/Player"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/Player"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "delete": {
        "tags": [
          "Players"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/Player"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Player"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Player"
                }
              }
            }
          }
        }
      }
    },
    "/api/PlayerName/{id}": {
      "get": {
        "tags": [
          "Players"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              },
              "application/json": {
                "schema": {
                  "type": "string"
                }
              },
              "text/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "Country": {
        "type": "object",
        "properties": {
          "Id": {
            "type": "integer",
            "format": "int64"
          },
          "Code": {
            "type": "string",
            "nullable": true
          },
          "Name": {
            "type": "string",
            "nullable": true
          },
          "ImageLink": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "Player": {
        "type": "object",
        "properties": {
          "Id": {
            "type": "integer",
            "format": "int64"
          },
          "Name": {
            "type": "string",
            "nullable": true
          },
          "CountryId": {
            "type": "integer",
            "format": "int64"
          },
          "Handed": {
            "type": "string",
            "nullable": true
          },
          "Dob": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "HomeTown": {
            "type": "string",
            "nullable": true
          },
          "HeightFeet": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "HeightInches": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "Weight": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "Photo": {
            "type": "string",
            "format": "byte",
            "nullable": true
          },
          "Gender": {
            "type": "string"
          },
          "TurnedPro": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "Country": {
            "$ref": "#/components/schemas/Country"
          }
        },
        "additionalProperties": false
      },
...
    }
  }
}